% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seerac.ellis.R
\name{getDebug}
\alias{getDebug}
\title{Get the current debug messaging level}
\usage{
getDebug()
}
\value{
The current debug messaging level, 0 for off, 1 for basic, 2 for advanced.
}
\description{
Get the current debug messaging level
}
