% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seerac.ellis.R
\name{getMaxGPUMemory}
\alias{getMaxGPUMemory}
\title{Get the total memory capacity of the GPU}
\usage{
getMaxGPUMemory()
}
\value{
The maximum memory of the GPU, NULL on failure
}
\description{
Get the total memory capacity of the GPU
}
