% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seerac.ellis.R
\docType{package}
\name{seerac.ellis}
\alias{seerac.ellis}
\title{seerac.ellis: Benthic Recovery Algorithm for the SEERAC Project}
\usage{
seerac.ellis(offsets, lookup, data)
}
\arguments{
\item{offsets}{integer vector. The indexing parameters of the data and lookup blocks.}

\item{lookup}{integer matrix. The lookup table which is loaded into GPU const memory.}

\item{data}{float matrix. Input data table which is used to produce an output results matrix}
}
\value{
NULL if an error occurred, A float matrix containing the results.
}
\description{
Benthic recovery algorithm for the SEERAC project utilising GPU CUDA. Based
off the Ellis et al paper.
}
\section{Mypackage functions}{

The mypackage functions ...
}

