% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seerac.ellis.R
\name{seerac.ellisWithFiles}
\alias{seerac.ellisWithFiles}
\title{Emulate the original implementation of the Benthic Recovery algorithm which saved all the input
data to files and then called a compiled executable. Assumes the caller has generated the input
files required. Used for testing purposes only.}
\usage{
seerac.ellisWithFiles(offsets)
}
\arguments{
\item{offsets}{integer vector. The indexing parameters of the data and lookup blocks.}
}
\value{
The result of the algorithm. 0 for success, negative number for error
}
\description{
Emulate the original implementation of the Benthic Recovery algorithm which saved all the input
data to files and then called a compiled executable. Assumes the caller has generated the input
files required. Used for testing purposes only.
}
