% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seerac.ellis.R
\name{setDebug}
\alias{setDebug}
\title{Set debug messaging level}
\usage{
setDebug(debug)
}
\arguments{
\item{debug}{integer, 0 for off, 1 for basic, 2 for advanced}
}
\description{
Set debug messaging level
}
